requirements.txt
deepface==0.0.89
streamlit==1.35.0
supabase==2.5.1
python-dotenv==1.0.1
vecs==0.4.3
opencv-python==4.9.0.80
tensorflow==2.16.1
tf_keras==2.16.0
ultralytics==8.2.79